spooky-c: clarify licensing of spooky-c by CC0 license file
I'm making an assumption (based on the comments in the .c file), that
Andi intended this code to be released to the public domain. Public
domain as a concept however is somewhat ill defined, especially when it
comes to different legal jurisdictions.

The creative commons folks however, have a license that is intended to
allow for the relenquishment of rights as much as possible in as many
jurisdictions as possible. It's called the CC0 public domain dedication.

I'd like to propose clarifying the licensing of spooky-c by putting it
under the CC0 license. This patch does that.

That said, if the intent was not to put this code into the public
domain, then please do let me know and I'll be happy to respin this
patch with the actual license intended.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>